apps/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
email-validator==2.1.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1
psycopg2-binary==2.9.11

# Redis & Background Jobs
redis==4.6.0
celery==5.3.4

# External Services
twilio==8.10.0
sendgrid==6.11.0
httpx==0.26.0
boto3==1.34.0

# Firebase Admin
firebase-admin==6.2.0

# AI/ML
groq==0.4.2
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
fastembed>=0.2.0
onnxruntime>=1.14.0

# PDF Processing
PyMuPDF==1.23.8
PyPDF2==3.0.1
Pillow==10.2.0
pdf2image==1.16.3

# Utilities
python-dotenv==1.0.0
apscheduler==3.10.4

# WebSockets
websockets==12.0
python-socketio==5.10.0
